# hst-server/swagger/swagger.yaml
basePath: /
definitions:
  hstserver_pkg_http.RetCode:
    enum:
    - 0
    - 1000
    - 1001
    - 1002
    - 1011
    - 1012
    - 1018
    - 1020
    - 1024
    - 1026
    - 1027
    - 1034
    - 60001
    - 60002
    type: integer
    x-enum-comments:
      RetAccountLocked: too many failed attempts, temporarily locked
      RetAuthAccountDisabled: the account is disabled
      RetAuthAccountInvalid: invalid account, our wrong password case
      RetAuthAccountUnknown: unknown account
      RetAuthApiDisabled: api connection prohibited, USER_RIGHT_API_ENABLED
      RetAuthClientInvalid: invalid type of the terminal
      RetAuthManagerIpBlock: ip address is not valid for the manager
      RetAuthManagerNoConfig: no manager configuration for the account
      RetAuthManagerType: this connection type is not permitted for manager
      RetAuthOtpInvalid: invalid one time password
      RetAuthResetPassword: master password must be changed
      RetAuthServerBusy: the server is busy
      RetSessionExpired: the session is gone, refresh or log in again
    x-enum-descriptions:
    - ""
    - invalid type of the terminal
    - invalid account, our wrong password case
    - the account is disabled
    - no manager configuration for the account
    - ip address is not valid for the manager
    - the server is busy
    - unknown account
    - this connection type is not permitted for manager
    - master password must be changed
    - invalid one time password
    - api connection prohibited, USER_RIGHT_API_ENABLED
    - the session is gone, refresh or log in again
    - too many failed attempts, temporarily locked
    x-enum-varnames:
    - RetOK
    - RetAuthClientInvalid
    - RetAuthAccountInvalid
    - RetAuthAccountDisabled
    - RetAuthManagerNoConfig
    - RetAuthManagerIpBlock
    - RetAuthServerBusy
    - RetAuthAccountUnknown
    - RetAuthManagerType
    - RetAuthResetPassword
    - RetAuthOtpInvalid
    - RetAuthApiDisabled
    - RetSessionExpired
    - RetAccountLocked
  internal_server_v1.ErrorResponse:
    properties:
      code:
        $ref: '#/definitions/hstserver_pkg_http.RetCode'
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_server_v1.ViewManagerRights:
    properties:
      groups:
        items:
          type: string
        type: array
      login:
        type: integer
      name:
        type: string
      rights:
        additionalProperties:
          type: boolean
        type: object
    type: object
...
info:
  contact:
    email: support@hybridsolutions.com
    name: API Support
  description: HS Trader APIs doc
  license:
    name: HS Trader
    url: https://hybridsolutions.com
  termsOfService: https://hybridsolutions.com
  title: HS Trader
  version: "1.0"
paths:
...
  /api/v1/managers:
    get:
      operationId: ListManagers
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_v1.ErrorResponse'
      security:
      - BearerAuth: []
      tags:
      - Managers
  /api/v1/managers/{login}/rights:
    get:
      operationId: GetManagerRights
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_v1.ViewManagerRights'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_v1.ErrorResponse'
      security:
      - BearerAuth: []
      tags:
      - Managers
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"